refactor: 커스텀 핸들러 구현부와 was 핸들러 구현부 분리
