+++ tomcat/src/main/java/com/techcourse/handler/RegisterHandler.java
@@ -1,7 +1,8 @@
-package org.apache.catalina;
+package com.techcourse.handler;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
+import org.apache.catalina.handler.Handler;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
