+++ tomcat/src/main/java/org/apache/catalina/handler/FileHandler.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.catalina.handler;
 
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
